+++ components/alarmdecoder-api/alarmdecoder_api.cpp
@@ -23,8 +23,12 @@
  */
 
 #include "alarmdecoder_api.h"
-
+#if defined(IDF_VER)
+#include "esp_log.h"
 static const char *TAG = "AD2API";
+#endif
+
+#define ZONE_TIMEOUT 60
 
 // nostate
 AD2VirtualPartitionState *nostate = nullptr;
@@ -58,7 +62,9 @@ void AlarmDecoderParser::reset_parser()
 void AlarmDecoderParser::updateVersion(char *arg)
 {
     std::string version = arg;
+#if defined(IDF_VER)
     ESP_LOGI(TAG, "updateVersion %s",arg);
+#endif
     notifySubscribers(ON_FIRMWARE_VERSION, version, nullptr);
 }
 
@@ -121,16 +127,105 @@ std::string AlarmDecoderParser::hex_to_binsz(void const * const ptr)
     return out;
 }
 
+/**
+ * @brief find value by name in query string config data
+ *    param1=val1&param2=val2
+ *
+ * @arg [in]qry_string std::string * to scan for NV data.
+ * @arg [in]key char * key to search for.
+ * @arg [in]val std::string &. Result buffer for value.
+ * @arg [in]val_size size of output buffer.
+ *
+ * @return int results. < -2 error, -1 not found, >= 0 result length
+ *
+ * @note val is cleared after param check.
+ */
+int AlarmDecoderParser::query_key_value_string(std::string  &qry_str, const char *key, std::string &val)
+{
+
+    /* Test parmeters. */
+    if ( !qry_str.length() || key == NULL) {
+        return -2;
+    }
+
+    /* Clear val first. */
+    val = "";
+
+    /* Init state machine args and get raw pointer to our query string. */
+    int keylen = strlen(key);
+    const char * qry_ptr = qry_str.c_str();
+
+    /* Process until we reach null terminator on the query string. */
+    while ( *qry_ptr ) {
+        const char *tp = qry_ptr;
+        int len = 0;
+        static char chr;
+
+        /* Scan the KEY looking for the next terminator. */
+        while ( (chr = *tp) != 0 ) {
+            if (chr == '=' || chr == '&') {
+                break;
+            }
+            len++;
+            tp++;
+        }
+
+        /* Test key for a match. */
+        if ( len && len == keylen ) {
+            if ( strncasecmp (key, qry_ptr, keylen) == 0 ) {
+
+                /* move the index */
+                len++;
+                tp++;
+
+                /* Test for null value. Still valid just return empty string. */
+                if ( !chr || chr == '&' ) {
+                    return val.length();
+                }
+
+                /* Save the value. */
+                while ( ( chr = *tp ) != 0 ) {
+                    if ( chr == '=' || chr == '&' ) {
+                        break;
+                    }
+                    val += chr;
+                    tp++;
+                }
+                return val.length();
+            }
+        }
+
+        /* End of string and key not found. We are done. */
+        if ( !chr ) {
+            return -1;
+        }
+
+        /* Keep looking skip last terminator. */
+        len++;
+
+        /* Scan till we start the next set. */
+        qry_ptr += len;
+        while ( chr && chr != '&' ) {
+            chr = *qry_ptr++;
+        }
+
+        /* End of string and not found. We are done. */
+        if ( !chr ) {
+            return -1;
+        }
+    }
+    return -1;
+}
 
 /**
  * @brief Subscribe to a EVENT type.
  *
  * @param [in]ev ad2_event_t event TYPE
- *   ex. ON_MESSAGE
+ *   ex. ON_ALPHA_MESSAGE
  * @param [in]fn Callback pointer function type AD2ParserCallback_sub_t.
  * @param [in]arg pointer to argument to pass to subscriber on event.
  */
-void AlarmDecoderParser::subscribeTo(ad2_event_t ev, AD2ParserCallback_sub_t fn, void *arg)
+void AlarmDecoderParser::subscribeTo(ad2_event_t ev, AD2SubScriber::AD2ParserCallback_sub_t fn, void *arg)
 {
     subscribers_t& v = AD2Subscribers[ev];
     v.push_back(AD2SubScriber(fn, arg));
@@ -142,7 +237,7 @@ void AlarmDecoderParser::subscribeTo(ad2_event_t ev, AD2ParserCallback_sub_t fn,
  * @param [in]fn Callback pointer function type AD2ParserCallbackRawRXData_sub_t.
  * @param [in]arg pointer to argument to pass to subscriber on event.
  */
-void AlarmDecoderParser::subscribeTo(AD2ParserCallbackRawRXData_sub_t fn, void *arg)
+void AlarmDecoderParser::subscribeTo(AD2SubScriber::AD2ParserCallbackRawRXData_sub_t fn, void *arg)
 {
     subscribers_t& v = AD2Subscribers[ON_RAW_RX_DATA];
     v.push_back(AD2SubScriber(fn, arg));
@@ -159,7 +254,7 @@ void AlarmDecoderParser::notifyRawDataSubscribers(uint8_t *data, size_t len)
 {
     // notify any direct subscribers to this event type(ON_RAW_RX_DATA).
     for ( subscribers_t::iterator i = AD2Subscribers[ON_RAW_RX_DATA].begin(); i != AD2Subscribers[ON_RAW_RX_DATA].end(); ++i ) {
-        ((AD2ParserCallbackRawRXData_sub_t)i->fn)(data, len, i->varg);
+        ((AD2SubScriber::AD2ParserCallbackRawRXData_sub_t)i->fn)(data, len, i->varg);
     }
 }
 
@@ -173,11 +268,6 @@ void AlarmDecoderParser::notifyRawDataSubscribers(uint8_t *data, size_t len)
  */
 void AlarmDecoderParser::notifySubscribers(ad2_event_t ev, std::string &msg, AD2VirtualPartitionState *pstate)
 {
-    // notify any direct subscribers to this event type.
-    for ( subscribers_t::iterator i = AD2Subscribers[ev].begin(); i != AD2Subscribers[ev].end(); ++i ) {
-        ((AD2ParserCallback_sub_t)i->fn)(&msg, pstate, i->varg);
-    }
-
     // Build a human readable string of the event and send
     // as the msg to notifySearchSubscribers with a message type of "EVENT".
     // notifySearchSubscribers will check all search subscribers and look
@@ -192,58 +282,112 @@ void AlarmDecoderParser::notifySubscribers(ad2_event_t ev, std::string &msg, AD2
     }
 
     // build a simple event string that can be used by search.
-    if (pstate) {
-        switch ((int)ev) {
-        case ON_DISARM:
-            break;
-        case ON_ARM:
+    switch ((int)ev) {
+    case ON_DISARM:
+        break;
+    case ON_ARM:
+        if (pstate) {
             if (pstate->armed_stay) {
                 emsg += " STAY";
             }
             if (pstate->armed_away) {
                 emsg += " AWAY";
             }
-            break;
-        case ON_POWER_CHANGE:
+        }
+        break;
+    case ON_POWER_CHANGE:
+        if(pstate) {
             if(pstate->ac_power) {
                 emsg += " AC";
             } else {
                 emsg += " BATTERY";
             }
-            break;
-        case ON_READY_CHANGE:
+        }
+        break;
+    case ON_READY_CHANGE:
+        if(pstate) {
             if(!pstate->ready) {
                 emsg += " ON";
             } else {
                 emsg += " OFF";
             }
-            break;
-        case ON_ALARM_CHANGE:
+        }
+        break;
+    case ON_ALARM_CHANGE:
+        if(pstate) {
             if(pstate->alarm_sounding) {
                 emsg += " ON";
             } else {
                 emsg += " OFF";
             }
-            break;
-        case ON_FIRE:
+        }
+        break;
+    case ON_FIRE_CHANGE:
+        if(pstate) {
             if(pstate->fire_alarm) {
                 emsg += " ON";
             } else {
                 emsg += " OFF";
             }
-            break;
-        case ON_CHIME_CHANGE:
+        }
+        break;
+    case ON_CHIME_CHANGE:
+        if(pstate) {
             if(pstate->chime_on) {
                 emsg += " ON";
             } else {
                 emsg += " OFF";
             }
-            break;
-        default:
-            emsg += " " + msg;
         }
-
+        break;
+    case ON_EXIT_CHANGE:
+        if(pstate) {
+            if(pstate->exit_now) {
+                emsg += " ON";
+            } else {
+                emsg += " OFF";
+            }
+        }
+        break;
+    case ON_PROGRAMMING_CHANGE:
+        if(pstate) {
+            if(pstate->programming) {
+                emsg += " ON";
+            } else {
+                emsg += " OFF";
+            }
+        }
+        break;
+    case ON_ZONE_CHANGE:
+        // lookup the zone state.
+        if(pstate) {
+            if(pstate->zone_states[pstate->zone].state() == AD2_STATE_TROUBLE) {
+                emsg += " TROUBLE ";
+            } else if(pstate->zone_states[pstate->zone].state() == AD2_STATE_OPEN) {
+                emsg += " OPEN ";
+            } else if(pstate->zone_states[pstate->zone].state() == AD2_STATE_CLOSED) {
+                emsg += " CLOSE ";
+            }
+            // zero pad 3 digit zone number string
+            char zstr[4];
+            snprintf(zstr, sizeof(zstr), "%03d", (int)pstate->zone);
+            emsg += zstr;
+        }
+        break;
+    default:
+        emsg += " " + msg;
     }
+
+    // save results of human readable event message.
+    if (pstate) {
+        pstate->last_event_message = emsg;
+    }
+
+    // notify any direct subscribers to this event type.
+    for ( subscribers_t::iterator i = AD2Subscribers[ev].begin(); i != AD2Subscribers[ev].end(); ++i ) {
+        ((AD2SubScriber::AD2ParserCallback_sub_t)i->fn)(&msg, pstate, i->varg);
+    }
+
     // notify any search subscribers that are watching for the "EVENT" type. Provide
     // a human readable event description.
     // TODO: Document event messages
@@ -257,7 +401,7 @@ void AlarmDecoderParser::notifySubscribers(ad2_event_t ev, std::string &msg, AD2
  * @param [in]regex_search regex search structure.
  * @param [in]arg pointer to argument to pass to subscriber on event.
  */
-void AlarmDecoderParser::subscribeTo(AD2ParserCallback_sub_t fn, AD2EventSearch *event_search)
+void AlarmDecoderParser::subscribeTo(AD2SubScriber::AD2ParserCallback_sub_t fn, AD2EventSearch *event_search)
 {
     subscribers_t& v = AD2Subscribers[ON_SEARCH_MATCH];
     v.push_back(AD2SubScriber(fn, event_search));
@@ -358,10 +502,10 @@ void AlarmDecoderParser::notifySearchSubscribers(ad2_message_t mt, std::string &
                 }
             }
 
-            // Test FAULT REGEX list stop on first matching statement.
+            // Test TROUBLE REGEX list stop on first matching statement.
             /// only test if a list is supplied.
-            if (!done && eSearch->FAULT_REGEX_LIST.size()) {
-                for(auto &regexstr : eSearch->FAULT_REGEX_LIST) {
+            if (!done && eSearch->TROUBLE_REGEX_LIST.size()) {
+                for(auto &regexstr : eSearch->TROUBLE_REGEX_LIST) {
                     std::smatch m;
                     std::regex e(regexstr);
                     const auto& tmsg = msg;
@@ -369,8 +513,8 @@ void AlarmDecoderParser::notifySearchSubscribers(ad2_message_t mt, std::string &
                     bool res = std::regex_search(tmsg, m, e);
                     if (res) {
                         // no match next subscriber.
-                        outformat = eSearch->FAULT_OUTPUT_FORMAT;
-                        eSearch->setState(AD2_STATE_FAULT);
+                        outformat = eSearch->TROUBLE_OUTPUT_FORMAT;
+                        eSearch->setState(AD2_STATE_TROUBLE);
                         // Clear last output results before we collect new.
                         eSearch->RESULT_GROUPS.clear();
                         // save the regex group results if any.
@@ -387,7 +531,7 @@ void AlarmDecoderParser::notifySearchSubscribers(ad2_message_t mt, std::string &
             if (savedstate != eSearch->getState()) {
                 eSearch->last_message = msg;
                 eSearch->out_message = outformat; //FIXME do the formatting macro magic stuff.
-                ((AD2ParserCallback_sub_t)i->fn)(&msg, pstate, i->varg);
+                ((AD2SubScriber::AD2ParserCallback_sub_t)i->fn)(&msg, pstate, i->varg);
             }
 
             // All done with this subscriber. Next.
@@ -396,7 +540,8 @@ void AlarmDecoderParser::notifySearchSubscribers(ad2_message_t mt, std::string &
 }
 
 /**
- * @brief Return a partition state structure by 8bit keypad address 0-31(Ademco) or partition #(DSC).
+ * @brief Return a partition state structure by 8bit keypad address 0-31(Ademco) or partition # 1-8(DSC).
+ * 0 is reserved for system partition.
  *
  * @param [in]address 8bit signed in partition or keypad address.
  * @param [in]update if true update mask adding new bits.
@@ -411,6 +556,8 @@ AD2VirtualPartitionState * AlarmDecoderParser::getAD2PState(int address, bool up
 
 /**
  * @brief Return a partition state structure by 32bit keypad/partition mask.
+ * LSB is address 1 on Ademco & partition 1 on DSC.
+ * Mask 0x00000000 is reserved for system partition state.
  *
  * @param [in]amask pointer to 32bit mask to search for.
  * @param [in]update if true update mask adding new bits.
@@ -461,6 +608,42 @@ AD2VirtualPartitionState * AlarmDecoderParser::getAD2PState(uint32_t *amask, boo
     }
     return ad2ps;
 }
+
+
+/**
+ * @brief Return a alpha description of a zone state. Use the AD2ZoneAlpha string if found
+ * or the standard 'ZONE XXX' template if not.
+ *
+ * @param [in]zone int value for the zone from 1-N
+ * @param [out]alpha std::string * alpha for the zone.
+ *
+ */
+void AlarmDecoderParser::getZoneString(uint8_t zone, std::string &alpha)
+{
+    if (AD2ZoneAlpha.count(zone)) {
+        alpha = AD2ZoneAlpha[zone];
+    } else {
+        alpha = "ZONE ";
+        // zero pad 3 digit zone number string
+        char zstr[4];
+        snprintf(zstr, sizeof(zstr), "%03d", (int)zone);
+        alpha += zstr;
+    }
+}
+
+/**
+ * @brief Set the alpha description of a zone.
+ *
+ * @param [in]zone int value for the zone from 1-N
+ *
+ * @return const char * alpha for the zone.
+ *
+ */
+void AlarmDecoderParser::setZoneString(uint8_t zone, const char* alpha)
+{
+    AD2ZoneAlpha[zone] = alpha;
+}
+
 
 /**
  * @brief Consume bytes from an AlarmDecoder stream into a small ring
@@ -569,7 +752,7 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
             // Process full messages on CR or LF
             if ( ch == '\n' || ch == '\r') {
 
-#if MONITOR_PARSER_TIMING
+#if MONITOR_PARSER_TIMING && defined(IDF_VER)
                 // monitor processing time.
                 int64_t xStart, xEnd, xDifference;
                 xStart = esp_timer_get_time();
@@ -615,6 +798,52 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
                     } else if (msg.find("!EXP:") == 0) {
                         MESSAGE_TYPE = EXP_MESSAGE_TYPE;
                         notifySubscribers(ON_EXP, msg, nostate);
+                        // DSC Zone Tracking use EXP messages and convert to zones.
+                        if (panel_type == 'D') {
+                            uint8_t exp_addr = atoi(msg.substr(5,2).c_str());
+                            uint8_t exp_chan = atoi(msg.substr(8,2).c_str());
+                            uint8_t zone = (exp_addr * 8) + exp_chan;
+                            uint8_t value = atoi(msg.substr(11,2).c_str());
+
+                            // default to nostate object.
+                            ad2ps = nostate;
+
+                            // Find the state based upon the zone.
+                            std::map<uint32_t, AD2VirtualPartitionState *>::iterator vpart_it = AD2PStates.begin();
+
+                            // Look at each virtual partition for a zone list match
+                            // and send a notification for every matching partition.
+                            bool _zone_found = false;
+                            while (vpart_it != AD2PStates.end()) {
+                                // If zone is in the known zone list then use this virtual partition.
+                                std::list<uint8_t> *zl = &vpart_it->second->zone_list;
+                                if ( find (zl->begin(), zl->end(), zone) != zl->end()) {
+                                    _zone_found = true;
+                                    // Found a match. Get pointer to partition state that matches this zone
+                                    ad2ps = vpart_it->second;
+                                    // Update the zone state object No timeout needed for DSC
+                                    ad2ps->zone_states[zone].state(value > 0 ? AD2_STATE_OPEN : AD2_STATE_CLOSED);
+                                    // Set the effected zone for the partition state.
+                                    ad2ps->zone = zone;
+                                    // Send zone change notification with partition state if found
+                                    notifySubscribers(ON_ZONE_CHANGE, msg, ad2ps);
+                                    // Done. Zone can only be mapped to one partition.
+                                    break;
+                                }
+                                vpart_it++;
+                            }
+                            // If not found then use default system partition for state storage.
+                            if (!_zone_found) {
+                                uint32_t amask = 0;
+                                ad2ps = getAD2PState(&amask, true);
+                                // Update the zone state object No timeout needed for DSC
+                                ad2ps->zone_states[zone].state(value > 0 ? AD2_STATE_OPEN : AD2_STATE_CLOSED);
+                                // Set the effected zone for the partition state.
+                                ad2ps->zone = zone;
+                                // Send zone change notification with partition state if found
+                                notifySubscribers(ON_ZONE_CHANGE, msg, ad2ps);
+                            }
+                        }
                     } else if (msg.find("!RFX:") == 0) {
                         MESSAGE_TYPE = RFX_MESSAGE_TYPE;
                         // Expand the HEX value to a bit string for easy pattern matching.
@@ -648,7 +877,8 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
                         MESSAGE_TYPE = CRC_MESSAGE_TYPE;
                         notifySubscribers(ON_CRC, msg, nostate);
                     } else if (msg.find("!VER:") == 0) {
-                        // Parse the version string.
+                        // save the AlarmDecoder firmware version string.
+                        ad2_version_string = msg.substr(5);
                         // call ON_VER callback if enabled.
                         MESSAGE_TYPE = VER_MESSAGE_TYPE;
                         notifySubscribers(ON_VER, msg, nostate);
@@ -656,6 +886,17 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
                         // call ON_ERR callback if enabled.
                         MESSAGE_TYPE = ERR_MESSAGE_TYPE;
                         notifySubscribers(ON_ERR, msg, nostate);
+                    } else if (msg.find("!CONFIG>") == 0) {
+                        // save the AlarmDecoder firmware configuration string.
+                        ad2_config_string = msg.substr(8);
+                        // Early update AlarmDecoder panel mode.
+                        std::string mode;
+                        if (query_key_value_string(ad2_config_string, "MODE", mode) >= 0 ) {
+                            panel_type = mode[0];
+                        }
+                        // call ON_CFG callback if enabled.
+                        MESSAGE_TYPE = CFG_MESSAGE_TYPE;
+                        notifySubscribers(ON_CFG, msg, nostate);
                     }
                 } else {
                     // http://www.alarmdecoder.com/wiki/index.php/Protocol#Keypad
@@ -670,12 +911,9 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
                             // to use it as a storage key for the state.
                             uint32_t amask = strtol(msg.substr(AMASK_START, AMASK_END-AMASK_START).c_str(), nullptr, 16);
 
+                            // Convert to host order LSB is address 1 on Ademco & partition 1 on DSC
+                            // 0x00000000 is reserved for system partition state.
                             amask = AD2_NTOHL(amask);
-                            // Ademco/DSC: MASK 00000000 = System
-                            // Ademco 00000001 is keypad address 0
-                            // Ademco 00000002 is keypad address 1
-                            // DSC    00000002 is partition 1
-                            // Ademco 40000000 is keypad address 30
 
                             // Create or return a pointer to our partition storage class.
                             ad2ps = getAD2PState(&amask, true);
@@ -693,11 +931,18 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
                             // get the panel type first
                             ad2ps->panel_type = msg[PANEL_TYPE_BYTE];
 
-                            // triggers
+                            // Update the parser panel mode.
+                            panel_type = ad2ps->panel_type;
+
+                            // Numeric field section #2 used in logic.
+                            std::string numeric_message = msg.substr(SECTION_2_START, 3);
+
+                            // event triggers
                             bool SEND_FIRE_CHANGE  = false;
                             bool SEND_READY_CHANGE = false;
                             bool SEND_ARMED_CHANGE = false;
                             bool SEND_CHIME_CHANGE = false;
+                            bool SEND_PROGRAMMING_CHANGE = false;
                             bool SEND_POWER_CHANGE = false;
                             bool SEND_BATTERY_CHANGE = false;
                             bool SEND_ALARM_CHANGE = false;
@@ -711,6 +956,7 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
                             bool ENTRY_DELAY = is_bit_set(ENTRYDELAY_BYTE, msg.c_str());
                             bool READY = is_bit_set(READY_BYTE, msg.c_str());
                             bool CHIME_ON = is_bit_set(CHIME_BYTE, msg.c_str());
+                            bool PROGRAMMING = is_bit_set(PROGMODE_BYTE, msg.c_str());
                             bool EXIT_NOW = false;
                             bool FIRE_ALARM = is_bit_set(FIRE_BYTE, msg.c_str());
                             bool AC_POWER = is_bit_set(ACPOWER_BYTE, msg.c_str());
@@ -718,6 +964,10 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
                             bool ALARM_BELL = is_bit_set(ALARM_BYTE, msg.c_str());
                             bool ALARM_STICKY = is_bit_set(ALARMSTICKY_BYTE, msg.c_str());
                             bool ZONE_BYPASSED = is_bit_set(BYPASS_BYTE, msg.c_str());
+                            uint8_t extra_sys_1 = (uint8_t) strtol(msg.substr(ADEMCO_EXTRA_SYSB1, 2).c_str(), 0, 16);
+                            uint8_t extra_sys_2 = (uint8_t) strtol(msg.substr(ADEMCO_EXTRA_SYSB2, 2).c_str(), 0, 16);
+                            uint8_t extra_sys_3 = (uint8_t) strtol(msg.substr(ADEMCO_EXTRA_SYSB3, 2).c_str(), 0, 16);
+                            uint8_t extra_sys_4 = (uint8_t) strtol(msg.substr(ADEMCO_EXTRA_SYSB4, 2).c_str(), 0, 16);
 
                             // Get section #4 alpha message and upper case for later searching
                             string ALPHAMSG = msg.substr(SECTION_4_START, 32);
@@ -733,6 +983,11 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
                                 // Skip fire bit on SYSTEM messages restore to current so we dont trip an event.
                                 if (ADEMCO_SYS_MESSAGE) {
                                     FIRE_ALARM = ad2ps->fire_alarm;
+                                } else {
+                                    // clear fire if READY
+                                    if (ad2ps->fire_alarm && READY) {
+                                        FIRE_ALARM = 0;
+                                    }
                                 }
 
                                 // Restore battery state only track when it is a system message.
@@ -809,6 +1064,11 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
                                     SEND_CHIME_CHANGE = true;
                                 }
 
+                                // programming state change send
+                                if ( ad2ps->programming != PROGRAMMING ) {
+                                    SEND_PROGRAMMING_CHANGE = true;
+                                }
+
                                 // ac_power state change send
                                 if ( ad2ps->ac_power != AC_POWER ) {
                                     SEND_POWER_CHANGE = true;
@@ -871,14 +1131,14 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
 
                             // Save states for non even tracked changes
                             ad2ps->backlight_on  = is_bit_set(BACKLIGHT_BYTE, msg.c_str());
-                            ad2ps->programming_mode = is_bit_set(PROGMODE_BYTE, msg.c_str());
+                            ad2ps->programming = is_bit_set(PROGMODE_BYTE, msg.c_str());
                             ad2ps->alarm_event_occurred = is_bit_set(ALARMSTICKY_BYTE, msg.c_str());
                             ad2ps->system_issue = is_bit_set(SYSISSUE_BYTE, msg.c_str());
                             ad2ps->system_specific = (uint8_t) (msg[SYSSPECIFIC_BYTE] - '0') & 0xff;
                             ad2ps->beeps = msg[BEEPMODE_BYTE] - '0';
 
                             // Extract the numeric value from section #2 HEX & DEC mix keep as string.
-                            ad2ps->last_numeric_message = msg.substr(SECTION_2_START, 3);
+                            ad2ps->last_numeric_message = numeric_message;
 
                             // Extract the 32 char Alpha message from section #4.
                             ad2ps->last_alpha_message = msg.substr(SECTION_4_START, 32);
@@ -889,20 +1149,83 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
 
 
                             // Debugging / testing output
+#if defined(IDF_VER)
                             ESP_LOGD(TAG, "!DBG: SSIZE(%i) PID(%i) MASK(%08X) Ready(%i) Armed[Away(%i) Stay(%i)] Bypassed(%i) Exit(%i)",
                                      AD2PStates.size(),ad2ps->partition,amask,ad2ps->ready,ad2ps->armed_away,ad2ps->armed_stay,ad2ps->zone_bypassed,ad2ps->exit_now);
+#endif
 
-                            // Call ON_MESSAGE callback if enabled.
-                            notifySubscribers(ON_MESSAGE, msg, ad2ps);
+                            // Call ON_ALPHA_MESSAGE callback if enabled.
+                            notifySubscribers(ON_ALPHA_MESSAGE, msg, ad2ps);
 
                             // Send event if FIRE state changed
                             if ( SEND_FIRE_CHANGE ) {
-                                notifySubscribers(ON_FIRE, msg, ad2ps);
+                                notifySubscribers(ON_FIRE_CHANGE, msg, ad2ps);
                             }
 
                             // Send event if ready state changed
                             if ( SEND_READY_CHANGE ) {
                                 notifySubscribers(ON_READY_CHANGE, msg, ad2ps);
+                            }
+
+                            // Update zone tracking if Ademco panel zone list report
+                            if (ad2ps->panel_type == ADEMCO_PANEL) {
+                                // Restore all faulted zones ON_READY.
+                                if (SEND_READY_CHANGE && ad2ps->ready) {
+                                    for (std::pair<uint8_t, AD2ZoneState> e : ad2ps->zone_states) {
+                                        // If zone(e.first) is currently OPEN then CLOSE it and notify subscribers.
+                                        if (ad2ps->zone_states[e.first].state() != AD2_STATE_CLOSED) {
+                                            // Update the zone state object and set timeout
+                                            ad2ps->zone_states[e.first].state(AD2_STATE_CLOSED, monotonicTime()+ZONE_TIMEOUT);
+                                            // Set the effected zone for the partition state.
+                                            ad2ps->zone = e.first;
+                                            // Send zone change notification with partition state if found
+                                            notifySubscribers(ON_ZONE_CHANGE, msg, ad2ps);
+                                        }
+                                    }
+                                } else
+                                    // If _not_ ready and _not_ system message update the zone state.
+                                    if (!ADEMCO_SYS_MESSAGE && ad2ps->system_specific == 0 && extra_sys_4 != 0xff) {
+                                        // get the numeric section and use as a zone #. TODO: Test of HEX and convert to base 10.
+                                        uint8_t _zone = (uint8_t) strtol(ad2ps->last_numeric_message.c_str(), 0, 10);
+                                        bool _send_event = false;
+
+                                        // this message is part of the zone low battery report
+                                        // [00000011000100000A--],023,[f70600ef1023004018020000000000],"LOBAT 23                        "
+                                        if (ad2ps->battery_low) {
+                                            // Update the low_battery object and set timeout
+                                            if (ad2ps->zone_states[_zone].low_battery() == false) {
+                                                _send_event = true;
+                                            }
+                                            ad2ps->zone_states[_zone].low_battery(monotonicTime()+ZONE_TIMEOUT);
+                                        } else {
+                                            // standard zone fault report.
+                                            // [00000011000000000A--],002,[f70600ef1002000018020000000000],"FAULT 02                        "
+                                            // check zone(system_issue) set for zone fault report entry.
+                                            // [00000401000000100A--],009,[f700001f1009040208020000000000],"CHECK 09                        "
+                                            if (ad2ps->system_issue) {
+                                                // Update the zone state object and set timeout
+                                                if (ad2ps->zone_states[_zone].state() != AD2_STATE_TROUBLE) {
+                                                    _send_event = true;
+                                                }
+                                                ad2ps->zone_states[_zone].state(AD2_STATE_TROUBLE, monotonicTime()+ZONE_TIMEOUT);
+                                            } else {
+                                                // Update the zone state object and set timeout
+                                                if (ad2ps->zone_states[_zone].state() != AD2_STATE_OPEN) {
+                                                    _send_event = true;
+                                                }
+                                                ad2ps->zone_states[_zone].state(AD2_STATE_OPEN, monotonicTime()+ZONE_TIMEOUT);
+                                            }
+
+                                            // Set the effected zone for the partition state.
+                                            ad2ps->zone = _zone;
+
+                                            // Send event notification if needed.
+                                            if (_send_event) {
+                                                // Send zone change notification with partition state if found
+                                                notifySubscribers(ON_ZONE_CHANGE, msg, ad2ps);
+                                            }
+                                        }
+                                    }
                             }
 
                             // Send armed/disarm event
@@ -917,6 +1240,11 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
                             // Send event if chime_on state changed
                             if ( SEND_CHIME_CHANGE ) {
                                 notifySubscribers(ON_CHIME_CHANGE, msg, ad2ps);
+                            }
+
+                            // Send event if programming state changed
+                            if ( SEND_PROGRAMMING_CHANGE ) {
+                                notifySubscribers(ON_PROGRAMMING_CHANGE, msg, ad2ps);
                             }
 
                             // Send event if ac_power state changed
@@ -947,14 +1275,22 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
                         }
                     } else {
                         //TODO: Error statistics tracking
+#if defined(IDF_VER)
                         ESP_LOGE(TAG, "!ERR: BAD PROTOCOL PREFIX.");
+#endif
                     }
                 }
 
                 // call Search callback subscribers if a match is found for this message type.
                 notifySearchSubscribers(MESSAGE_TYPE, msg, ad2ps);
 
-#ifdef MONITOR_PARSER_TIMING
+                // Zone tracking timeouts.
+                // TODO: Add to external periodic call. If we dont get messages this wont run.
+                // Not a problem in most cases but in some cases such as DEDUPLICATE setting on the AD2*
+                // this could be a problem.
+                checkZoneTimeout();
+
+#ifdef MONITOR_PARSER_TIMING && defined(IDF_VER)
                 xEnd = esp_timer_get_time();
                 xDifference = xEnd - xStart;
                 ESP_LOGI(TAG, "message processing time: %lldus", xDifference );
@@ -986,6 +1322,46 @@ bool AlarmDecoderParser::put(uint8_t *buff, int8_t len)
     }
 
     return true;
+}
+
+/**
+ * @brief Check zones for timeouts and send notifications.
+ */
+void AlarmDecoderParser::checkZoneTimeout()
+{
+    AD2VirtualPartitionState *ad2ps = nullptr;
+    // Find the state based upon the zone.
+    std::map<uint32_t, AD2VirtualPartitionState *>::iterator vpart_it = AD2PStates.begin();
+
+    // Look at each virtual partition zone list for state timeouts.
+    // and send a notification for every matching partition.
+    while (vpart_it != AD2PStates.end()) {
+        ad2ps = vpart_it->second;
+        if (ad2ps) {
+            std::string msg = "ZONE_CHECK";
+            for (std::pair<uint8_t, AD2ZoneState> e : ad2ps->zone_states) {
+                // If zone is OPEN and the reset time has expired restore and notify subscribers.
+                if (ad2ps->zone_states[e.first].state() != AD2_STATE_CLOSED) {
+                    unsigned long _reset_time = ad2ps->zone_states[e.first].state_reset_time();
+                    if (_reset_time && _reset_time < monotonicTime()) {
+                        ad2ps->zone_states[e.first].state(AD2_STATE_CLOSED);
+                        ad2ps->zone = e.first;
+                        notifySubscribers(ON_ZONE_CHANGE, msg, ad2ps);
+                    }
+                }
+                // If zone Battery is faulted and reset time has expired restore and notify subscribers.
+                if (ad2ps->zone_states[e.first].low_battery()) {
+                    unsigned long _reset_time = ad2ps->zone_states[e.first].battery_reset_time();
+                    if (_reset_time && _reset_time < monotonicTime()) {
+                        ad2ps->zone_states[e.first].low_battery(false);
+                        ad2ps->zone = e.first;
+                        notifySubscribers(ON_ZONE_CHANGE, msg, ad2ps);
+                    }
+                }
+            }
+        }
+        vpart_it++;
+    }
 }
 
 /**