+++ main/ad2_settings.h
@@ -32,6 +32,9 @@
 // @brief MAX virtual partition slots
 #define AD2_MAX_VPARTITION    9
 
+// @brief MAX zones
+#define AD2_MAX_ZONES         255
+
 // @brief MAX code slots
 #define AD2_MAX_CODE          99
 
@@ -46,6 +49,10 @@
 
 // @brief Parent Key to store virtual partition slots
 #define VPART_CONFIG_KEY     "vpa"
+#define VPART_ZL_CONFIG_KEY  "vpz"
+
+// @brief Parent Key to store zone slots
+#define ZONES_ALPHA_CONFIG_KEY      "zna"
 
 // @brief Parent Key to store code slots
 #define CODES_CONFIG_KEY      "codes"